feat: Split authentication into separate class file
